+++ src/server/metiers/infra/apiLaBonneAlternanceMétier.repository.test.ts
@@ -1,16 +1,12 @@
-import { Failure, Success } from '~/server/errors/either';
+import { createFailure, Failure, Success } from '~/server/errors/either';
 import { ErreurMétier } from '~/server/errors/erreurMétier.types';
 import { Métier } from '~/server/metiers/domain/métier';
 import { aListeDeMetierLaBonneAlternance } from '~/server/metiers/domain/métier.fixture';
-import {
-	aMetierLaBonneAlternanceApiResponse,
-} from '~/server/metiers/infra/apiLaBonneAlternanceMétier.fixture';
-import {
-	ApiLaBonneAlternanceMétierRepository,
-} from '~/server/metiers/infra/apiLaBonneAlternanceMétier.repository';
+import { aMetierLaBonneAlternanceApiResponse } from '~/server/metiers/infra/apiLaBonneAlternanceMétier.fixture';
+import { ApiLaBonneAlternanceMétierRepository } from '~/server/metiers/infra/apiLaBonneAlternanceMétier.repository';
+import { anErrorManagementService } from '~/server/services/error/errorManagement.fixture';
 import { anHttpError } from '~/server/services/http/httpError.fixture';
 import { anAxiosResponse, aPublicHttpClientService } from '~/server/services/http/publicHttpClient.service.fixture';
-import { aLoggerService } from '~/server/services/logger.service.fixture';
 
 describe('ApiLaBonneAlternanceMétierRepository', () => {
 	describe('getMetierList', () => {
@@ -19,7 +15,8 @@ describe('ApiLaBonneAlternanceMétierRepository', () => {
 				const httpClientService = aPublicHttpClientService();
 				(httpClientService.get as jest.Mock).mockResolvedValue(anAxiosResponse(aMetierLaBonneAlternanceApiResponse()));
 				const expected = aListeDeMetierLaBonneAlternance();
-				const repository = new ApiLaBonneAlternanceMétierRepository(httpClientService, aLoggerService());
+
+				const repository = new ApiLaBonneAlternanceMétierRepository(httpClientService, anErrorManagementService());
 
 				const response = await repository.getMetierList('tran');
 
@@ -31,15 +28,26 @@ describe('ApiLaBonneAlternanceMétierRepository', () => {
 
 		describe('Quand l‘api renvoie une erreur', () => {
 			it("retourne une instance d'erreur", async () => {
-				const httpClientService = aPublicHttpClientService();
-				(httpClientService.get as jest.Mock).mockRejectedValue(anHttpError(429));
-				const repository = new ApiLaBonneAlternanceMétierRepository(httpClientService, aLoggerService());
+				const httpError = anHttpError(429);
+				const httpClientService = aPublicHttpClientService({
+					get: jest.fn(async () => {
+						throw httpError;
+					}),
+				});
+				const expectedFailure = ErreurMétier.DEMANDE_INCORRECTE;
+				const errorManagementService = anErrorManagementService({ handleFailureError: jest.fn(() => createFailure(expectedFailure)) });
+				const repository = new ApiLaBonneAlternanceMétierRepository(httpClientService, errorManagementService);
 
 				const response = await repository.getMetierList('tran');
 
 				expect(httpClientService.get).toHaveBeenCalledTimes(1);
+				expect(errorManagementService.handleFailureError).toHaveBeenCalledWith(httpError, {
+					apiSource: 'API LaBonneAlternance',
+					contexte: 'get metier la bonne alternance',
+					message: '[API LaBonneAlternance] impossible de récuperer les métiers',
+				});
 				expect(response.instance).toEqual('failure');
-				expect((response as Failure).errorType).toEqual(ErreurMétier.DEMANDE_INCORRECTE);
+				expect((response as Failure).errorType).toEqual(expectedFailure);
 			});
 		});
 	});