+++ src/server/services/http/publicHttpClient.service.fixture.ts
@@ -73,12 +73,13 @@ export function anAxiosError(override?: Partial<AxiosError>): AxiosError {
 	} as unknown as AxiosError;
 }
 
-export function aPublicHttpClientService(): PublicHttpClientService {
+export function aPublicHttpClientService(override?: Partial<PublicHttpClientService> ): PublicHttpClientService {
 	return {
 		client: anAxiosInstance(),
 		get: jest.fn(),
 		post: jest.fn(),
 		setAuthorizationHeader: jest.fn(),
+		...override,
 	} as unknown as PublicHttpClientService;
 }